backend/requirements.txt
fastapi
uvicorn
pydantic
pytest
httpx
openai
requests
python-dotenv
tinydb
pymc==5.1.2
arviz==0.14.0
scipy==1.10.*
numpy==1.24.*

